requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0

# CLIP support
open-clip-torch>=2.20.0

# Configuration and logging
pyyaml>=6.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Testing
pytest>=7.3.0

# Optional: For reproducibility
scipy>=1.10.0
